Moves styleguide tests
