+++ packages/styleguide/__tests__/android/styleguide.android.test.js
